Enhancement: split out tests
