+++ pyproject.toml
@@ -15,7 +15,7 @@ Sphinx = ">=1.3"
 sphinx_rtd_theme = ">=0.3.1"
 behave = ">=1.2.6"
 recommonmark = ">=0.4.0"
-pytest-bdd = { version = "^3.2.1", optional = true }
+pytest-bdd = {git = "https://github.com/rbcasperson/pytest-bdd.git", rev = "scenario-descriptions", optional = true}
 
 [tool.poetry.dev-dependencies]
 pre-commit = "^1.15"
